+++ bins/revme/src/cmd/bench/snailtracer.rs
@@ -1,10 +1,11 @@
 use context::TxEnv;
 use criterion::Criterion;
 use database::{BenchmarkDB, BENCH_CALLER, BENCH_TARGET};
+use inspector::CountInspector;
 use revm::{
     bytecode::Bytecode,
     primitives::{bytes, hex, Bytes, TxKind},
-    Context, ExecuteEvm, MainBuilder, MainContext,
+    Context, ExecuteEvm, InspectEvm, MainBuilder, MainContext,
 };
 
 pub fn run(criterion: &mut Criterion) {
@@ -13,7 +14,8 @@ pub fn run(criterion: &mut Criterion) {
     let mut evm = Context::mainnet()
         .with_db(BenchmarkDB::new_bytecode(bytecode.clone()))
         .modify_cfg_chained(|c| c.disable_nonce_check = true)
-        .build_mainnet();
+        .build_mainnet()
+        .with_inspector(CountInspector::new());
 
     let tx = TxEnv::builder()
         .caller(BENCH_CALLER)
@@ -31,6 +33,19 @@ pub fn run(criterion: &mut Criterion) {
             },
             |input| {
                 let _ = evm.transact_one(input).unwrap();
+            },
+            criterion::BatchSize::SmallInput,
+        );
+    });
+
+    criterion.bench_function("analysis-inspector", |b| {
+        b.iter_batched(
+            || {
+                // create a transaction input
+                tx.clone()
+            },
+            |input| {
+                let _ = evm.inspect_one_tx(input);
             },
             criterion::BatchSize::SmallInput,
         );